ForwardProblem refactoring :factory:
